+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemDyeStack.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.item.interfaces;
 
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.item.component.ItemDyeComponent;
 
-public interface ItemDyeStack extends ItemStack {
+public interface ItemDyeStack extends ItemStack, ItemDyeComponent {
 }