+++ server/src/main/java/org/allaymc/server/container/processor/CreateActionProcessor.java
@@ -3,14 +3,14 @@ package org.allaymc.server.container.processor;
 import lombok.extern.slf4j.Slf4j;
 import org.allaymc.api.container.ContainerTypes;
 import org.allaymc.api.entity.interfaces.EntityPlayer;
-import org.allaymc.api.item.recipe.Recipe;
+import org.allaymc.api.item.ItemStack;
 import org.cloudburstmc.protocol.bedrock.data.inventory.itemstack.request.action.CreateAction;
 import org.cloudburstmc.protocol.bedrock.data.inventory.itemstack.request.action.ItemStackRequestAction;
 import org.cloudburstmc.protocol.bedrock.data.inventory.itemstack.request.action.ItemStackRequestActionType;
 
 import java.util.Map;
 
-import static org.allaymc.server.container.processor.CraftRecipeActionProcessor.RECIPE_DATA_KEY;
+import static org.allaymc.server.container.processor.CraftRecipeActionProcessor.RECIPE_OUTPUTS_DATA_KEY;
 
 /**
  * @author daoge_cmd
@@ -19,14 +19,14 @@ import static org.allaymc.server.container.processor.CraftRecipeActionProcessor.
 public class CreateActionProcessor implements ContainerActionProcessor<CreateAction> {
     @Override
     public ActionResponse handle(CreateAction action, EntityPlayer player, int currentActionIndex, ItemStackRequestAction[] actions, Map<String, Object> dataPool) {
-        var recipe = (Recipe) dataPool.get(RECIPE_DATA_KEY);
-        if (recipe == null) {
+        var outputs = (ItemStack[]) dataPool.get(RECIPE_OUTPUTS_DATA_KEY);
+        if (outputs == null) {
             log.warn("Recipe not found in data pool!");
             return error();
         }
 
         // No need to copy because when we get item from created output, we will copy it
-        var output = recipe.getOutputs()[action.getSlot()];
+        var output = outputs[action.getSlot()];
         player.getContainer(ContainerTypes.CREATED_OUTPUT).setItemStack(0, output, false);
         return null;
     }