+++ server/src/test/java/org/allaymc/api/utils/AllayNBTUtilsTest.java
@@ -12,33 +12,33 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 /**
  * @author daoge_cmd
  */
-class AllayNbtUtilsTest {
+class AllayNBTUtilsTest {
     static String NBT_BASE64_LE = "CgAAAQgAdGVzdEZsYWcBAA==";
     static String NBT_BASE64 = "CgAAAQAIdGVzdEZsYWcBAA==";
 
     @Test
     void testNbtToBase64() {
         var nbtMap = NbtMap.builder().putBoolean("testFlag", true).build();
-        var nbtB64 = AllayNbtUtils.nbtToBase64(nbtMap);
+        var nbtB64 = AllayNBTUtils.nbtToBase64(nbtMap);
         assertEquals(NBT_BASE64, nbtB64);
     }
 
     @Test
     void testNbtToBase64LE() {
         var nbtMap = NbtMap.builder().putBoolean("testFlag", true).build();
-        var nbtB64 = AllayNbtUtils.nbtToBase64LE(nbtMap);
+        var nbtB64 = AllayNBTUtils.nbtToBase64LE(nbtMap);
         assertEquals(NBT_BASE64_LE, nbtB64);
     }
 
     @Test
     void testBase64ToNbt() {
-        var nbtMap = AllayNbtUtils.base64ToNbt(NBT_BASE64);
+        var nbtMap = AllayNBTUtils.base64ToNbt(NBT_BASE64);
         assertTrue(nbtMap.getBoolean("testFlag"));
     }
 
     @Test
     void testBase64ToNbtLE() {
-        var nbtMap = AllayNbtUtils.base64ToNbtLE(NBT_BASE64_LE);
+        var nbtMap = AllayNBTUtils.base64ToNbtLE(NBT_BASE64_LE);
         assertTrue(nbtMap.getBoolean("testFlag"));
     }
 
@@ -48,19 +48,19 @@ class AllayNbtUtilsTest {
                 NbtMap.builder().putBoolean("testFlag1", true).build(),
                 NbtMap.builder().putBoolean("testFlag2", true).build()
         );
-        assertEquals(nbtList, AllayNbtUtils.bytesToNbtListLE(AllayNbtUtils.nbtListToBytesLE(nbtList)));
+        assertEquals(nbtList, AllayNBTUtils.bytesToNbtListLE(AllayNBTUtils.nbtListToBytesLE(nbtList)));
     }
 
     @Test
     void testBytesAndNbtConverter() {
         var nbt = NbtMap.builder().putBoolean("testFlag", true).build();
-        assertEquals(nbt, AllayNbtUtils.bytesToNbtLE(AllayNbtUtils.nbtToBytesLE(nbt)));
+        assertEquals(nbt, AllayNBTUtils.bytesToNbtLE(AllayNBTUtils.nbtToBytesLE(nbt)));
     }
 
     @Test
     void testWriteVector3f() {
         var nbt = NbtMap.builder();
-        AllayNbtUtils.writeVector3f(nbt, "test", new org.joml.Vector3f(1, 2, 3));
+        AllayNBTUtils.writeVector3f(nbt, "test", new org.joml.Vector3f(1, 2, 3));
         var nbtMap = nbt.build();
         var list = nbtMap.getList("test", NbtType.FLOAT);
         assertEquals(1, list.get(0));
@@ -72,7 +72,7 @@ class AllayNbtUtilsTest {
     void testReadVector3f() {
         var nbt = NbtMap.builder();
         nbt.putList("test", NbtType.FLOAT, 1f, 2f, 3f);
-        var vector3f = AllayNbtUtils.readVector3f(nbt.build(), "test");
+        var vector3f = AllayNBTUtils.readVector3f(nbt.build(), "test");
         assertEquals(1, vector3f.x);
         assertEquals(2, vector3f.y);
         assertEquals(3, vector3f.z);